GEOD-294 - Renamed one class to drop '.service'
