+++ src/client/app/shared/json-data-view-model/data-view-translator.ts
