+++ src/client/app/shared/json-data-view-model/index.ts
@@ -2,4 +2,4 @@
  * This barrel file provides the export for the shared SiteLogService.
  */
 export * from './json-view-model.service';
-export * from './data-view-translator.service';
+export * from './data-view-translator';
